Get storage context tests working
