+++ packages/client-ts/src/context/engines/verida/database/utils.ts
@@ -1,3 +1,5 @@
+import EncryptionUtils from '@verida/encryption-utils'
+
 /**
  * @category
  * Modules
@@ -12,6 +14,20 @@ class Utils {
     max = Math.floor(max);
     return Math.floor(Math.random() * (max - min) + min); // The maximum is exclusive and the minimum is inclusive
   }
+
+  // DID + context name + DB Name + readPerm + writePerm
+  static buildDatabaseHash(databaseName: string, contextName: string, did: string) {
+    let text = [
+        did.toLowerCase(),
+        contextName,
+        databaseName,
+    ].join("/");
+
+    const hash = EncryptionUtils.hash(text).substring(2);
+
+    // Database name in CouchDB must start with a letter, so prepend a `v`
+    return "v" + hash;
+}
 
   /**
    * Build an MD5 hash from an array
